updated files for working function
