+++ apply.html
@@ -14,7 +14,7 @@
     </header>
 
     <main>
-        <form id="name-application">
+        <form id="order-entry">
             <section>
                 <h2>About You</h2>
             
@@ -71,24 +71,24 @@
                     </legend>
 
                     <div>
-                        <input name="weapons" value="speakers" type="checkbox" id="features">
+                        <input name="features" value="speakers" type="checkbox" id="features">
                         <label for="speakers">Buil-In Bluetooth Speakers</label>
                     </div>
     
                     <div>
-                        <input name="weapons" value="head-warmer" type="checkbox" id="features">
+                        <input name="features" value="head-warmer" type="checkbox" id="features">
                         <label for="warmer">Head Warmer</label>
                     </div>
     
                     <div>
-                        <input name="weapons" value="head-massager" type="checkbox" id="features">
+                        <input name="features" value="head-massager" type="checkbox" id="features">
                         <label for="head-massager">Head Massager</label>
                     </div>
                 </fieldset>
             </section>
 
             <section>
-                <h2>Submit Entry</h2>
+                <h2>Submit Order</h2>
                 <button>Submit</button>
             </section>
         </form>
